fixing %c%s%i%u%d%%

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: aboudoun <aboudoun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 15:20:12 by aboudoun          #+#    #+#             */
-/*   Updated: 2021/12/15 18:40:54 by aboudoun         ###   ########.fr       */
+/*   Updated: 2021/12/16 13:27:34 by aboudoun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 int		ft_printf(const char *s, ... );
 void	ft_putchar(int c, int *len);
 void	ft_putstr(char *str, int *len);
-void	ft_putnbr(int nb , int *len);
+void	ft_putnbr(long nb , int *len);
 
 
 #endif
